Fix crash on corrupted selection change
